Add skeleton device auth card
