+++ src/lib/components/device-auth/SkeletonDeviceAuthCard.svelte
@@ -1,0 +1,35 @@
+<div class="border border-input rounded-md p-4.5 relative size-full bg-surface-alt skeleton-loader">
+  <div class="flex justify-between items-start">
+    <div class="flex flex-col gap-y-1">
+      <div class="h-6 w-32 rounded skeleton-loader mb-2"></div>
+
+      <div class="flex flex-col gap-y-4">
+        <div class="flex flex-col gap-1">
+          <div class="h-4 w-20 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+        </div>
+
+        <div class="flex flex-col gap-1">
+          <div class="h-4 w-20 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+        </div>
+
+        <div class="flex flex-col gap-2 mt-2">
+          <div class="h-4 w-24 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+        </div>
+
+        <div class="flex flex-col gap-2 mt-2">
+          <div class="h-4 w-24 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+          <div class="h-4 w-48 rounded skeleton-loader"></div>
+        </div>
+      </div>
+    </div>
+
+    <div class="absolute top-4 right-4 size-10 rounded skeleton-loader"></div>
+  </div>
+</div>
